+++ src/predicates/isDeprecatedMediaFeature/index.js
@@ -1,0 +1,5 @@
+import { cssDeprecatedMediaFeatures } from '../../constants/cssDeprecatedMediaFeatures';
+
+export function isDeprecatedMediaFeature(feature) {
+	return cssDeprecatedMediaFeatures.some((deprecatedFeature) => feature.endsWith(deprecatedFeature));
+}